+++ src/main/java/com/programmers/springbootbasic/service/dto/Voucher/VoucherCreationRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.springbootbasic.service.dto;
+package com.programmers.springbootbasic.service.dto.Voucher;
 
 import java.time.LocalDateTime;
